[WIP] new clang plugin: loopvartoosmall
Idea from bubli - look for loops where the index variable is of such
size that it cannot cover the range revealed by examining the length
part of the condition.

So far, I have only run the plugin up till the VCL module.

Also the plugin deliberately excludes anything more complicated than a
straightforward incrementing for loop.

+++ l10ntools/source/cfgmerge.cxx
@@ -470,7 +470,7 @@ void CfgMerge::WorkOnResourceEnd()
         if ( pEntrys ) {
             OString sCur;
 
-            for( unsigned int i = 0; i < aLanguages.size(); ++i ){
+            for( size_t i = 0; i < aLanguages.size(); ++i ){
                 sCur = aLanguages[ i ];
 
                 OString sContent;
